fix: show related document in CT Come Fare Per
